+++ python/examples/serve/a2a_server.py
@@ -8,7 +8,7 @@ from beeai_framework.tools.weather import OpenMeteoTool
 
 
 def main() -> None:
-    llm = ChatModel.from_name("ollama:granite3.3:8b")
+    llm = ChatModel.from_name("ollama:granite4:micro")
     agent = RequirementAgent(
         llm=llm,
         tools=[DuckDuckGoSearchTool(), OpenMeteoTool()],
@@ -16,7 +16,7 @@ def main() -> None:
     )
 
     # Register the agent with the A2A server and run the HTTP server
-    # For the ToolCallingAgent, we don't need to specify ACPAgent factory method
+    # For the ToolCallingAgent, we don't need to specify A2AAgent factory method
     # because it is already registered in the A2AServer
     # we use LRU memory manager to keep limited amount of sessions in the memory
     A2AServer(